style updates and database fixes
minor fixes, quality of life

+++ app/models.py
@@ -4,7 +4,7 @@ class Item(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     item_name = db.Column(db.String)
     category = db.Column(db.String, nullable=False)
-    price = db.Column(db.Float, nullable=False)
+    price = db.Column(db.Numeric, nullable=False)
 
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
